refactor (adapter) : handle 메소드를 통합한다

+++ src/main/java/kr/co/koscom/miniproject/stock/application/dto/response/RetrieveStockChartResponse.java
@@ -24,7 +24,7 @@ public record RetrieveStockChartResponse(
     record StockComponent(
         String stockName,
         Integer currentPrice,
-        Integer tradingVolume,
+        Long tradingVolume,
         Double fluctuationRate
     ) {
